+++ packages/core/src/models/index.ts
@@ -1,3 +1,2 @@
 export * from './devided-page-path';
 export * from './vo/error-apiv3';
-export * from './markdown-table';
